#18 :sparkles: Add HTML raise command
Also restructured TS-command code
